Macときのみスクショ用のディレクトリを配置する

+++ initialize.sh
@@ -35,7 +35,11 @@ cd $HOME
 # starshipのインストール
 sh -c "$(curl -fsSL https://starship.rs/install.sh)" -- --yes
 
-# スクリーンショットの保存先をDownloadにする
-mkdir -p ${HOME}/Downloads/.screencapture
-defaults write com.apple.screencapture location ${HOME}/Downloads/.screencapture
-killall SystemUIServer
+# 【Mac】スクリーンショットの保存先をDownloadにする
+if test -e /etc/os-release ; then
+  :
+else
+  mkdir -p ${HOME}/Downloads/.screencapture
+  defaults write com.apple.screencapture location ${HOME}/Downloads/.screencapture
+  killall SystemUIServer
+fi
